DbTrackRepository を廃止し、関数直接呼び出しに変更

+++ domain/src/folder/usecase.rs
@@ -9,7 +9,7 @@ use sqlx::PgTransaction;
 
 use super::FolderIdMayRoot;
 use crate::{
-    Error, folder::folder_repository, path::LibraryDirectoryPath, track::DbTrackRepository,
+    Error, folder::folder_repository, path::LibraryDirectoryPath, track::track_repository,
 };
 
 /// ライブラリのフォルダ関係のUsecase
@@ -28,18 +28,10 @@ pub trait FolderUsecase {
 
 /// FolderUsecaseの本実装
 #[derive(new)]
-pub struct FolderUsecaseImpl<SR>
-where
-    SR: DbTrackRepository + Sync + Send,
-{
-    db_track_repository: SR,
-}
+pub struct FolderUsecaseImpl {}
 
 #[async_trait]
-impl<SR> FolderUsecase for FolderUsecaseImpl<SR>
-where
-    SR: DbTrackRepository + Sync + Send,
-{
+impl FolderUsecase for FolderUsecaseImpl {
     /// フォルダに曲が含まれてない場合、削除する
     ///
     /// # Arguments
@@ -54,60 +46,45 @@ where
             .await?
             .ok_or_else(|| Error::DbFolderPathNotFound(folder_path.to_owned()))?;
 
-        self.delete_db_if_empty_by_id(tx, folder_id).await
+        delete_db_if_empty_by_id(tx, folder_id).await
     }
 }
 
-impl<SR> FolderUsecaseImpl<SR>
-where
-    SR: DbTrackRepository + Sync + Send,
-{
-    /// フォルダに曲が含まれてない場合、削除する(再帰実行用のID指定版)
-    ///
-    /// # Arguments
-    /// - folder_path: 確認・削除対象のフォルダパス
-    #[async_recursion]
-    async fn delete_db_if_empty_by_id<'c>(
-        &self,
-        tx: &mut PgTransaction<'c>,
-        folder_id: i32,
-    ) -> Result<()> {
-        //他の曲が含まれる場合、削除せずに終了
-        if self
-            .db_track_repository
-            .is_exist_in_folder(tx, folder_id)
-            .await?
-        {
+/// フォルダに曲が含まれてない場合、削除する(再帰実行用のID指定版)
+///
+/// # Arguments
+/// - folder_path: 確認・削除対象のフォルダパス
+#[async_recursion]
+async fn delete_db_if_empty_by_id<'c>(tx: &mut PgTransaction<'c>, folder_id: i32) -> Result<()> {
+    //他の曲が含まれる場合、削除せずに終了
+    if track_repository::is_exist_in_folder(tx, folder_id).await? {
+        return Ok(());
+    }
+
+    let parent_id_mr = {
+        //他のフォルダが含まれる場合、削除せずに終了
+        if folder_repository::is_exist_in_folder(tx, FolderIdMayRoot::Folder(folder_id)).await? {
             return Ok(());
         }
 
-        let parent_id_mr = {
-            //他のフォルダが含まれる場合、削除せずに終了
-            if folder_repository::is_exist_in_folder(tx, FolderIdMayRoot::Folder(folder_id)).await?
-            {
-                return Ok(());
-            }
+        //削除するフォルダ情報を取得
+        let parent_id_mr = folder_repository::get_parent(tx, folder_id)
+            .await?
+            .ok_or(Error::DbFolderIdNotFound(folder_id))?;
 
-            //削除するフォルダ情報を取得
-            let parent_id_mr = folder_repository::get_parent(tx, folder_id)
-                .await?
-                .ok_or(Error::DbFolderIdNotFound(folder_id))?;
+        //削除を実行
+        folder_repository::delete(tx, folder_id).await?;
 
-            //削除を実行
-            folder_repository::delete(tx, folder_id).await?;
+        parent_id_mr
+    };
 
-            parent_id_mr
-        };
-
-        //親フォルダについて再帰実行
-        if let FolderIdMayRoot::Folder(parent_id) = parent_id_mr {
-            self.delete_db_if_empty_by_id(tx, parent_id).await?;
-        }
-
-        Ok(())
+    //親フォルダについて再帰実行
+    if let FolderIdMayRoot::Folder(parent_id) = parent_id_mr {
+        delete_db_if_empty_by_id(tx, parent_id).await?;
     }
-}
 
+    Ok(())
+}
 #[derive(Default)]
 pub struct MockFolderUsecase {
     pub inner: MockFolderUsecaseInner,
